requirements.txt
Werkzeug==3.0.1
Flask==3.0.0
Flask-Cors==4.0.0
numpy==1.26.4
tensorflow-cpu==2.15.0
